+++ typescript/sdk/src/token/TokenStandard.ts
@@ -173,6 +173,9 @@ export const TOKEN_COLLATERALIZED_STANDARDS = [
   TokenStandard.EvmHypVSXERC20Lockbox,
   TokenStandard.AleoHypNative,
   TokenStandard.AleoHypCollateral,
+  TokenStandard.RadixHypCollateral,
+  TokenStandard.StarknetHypCollateral,
+  TokenStandard.StarknetHypNative,
 ];
 
 export const XERC20_STANDARDS = [