+++ typescript/radix-sdk/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @hyperlane-xyz/radix-sdk
 
+## 25.3.2
+
+### Patch Changes
+
+- @hyperlane-xyz/utils@25.3.2
+- @hyperlane-xyz/provider-sdk@1.3.6
+
+## 25.3.1
+
+### Patch Changes
+
+- @hyperlane-xyz/utils@25.3.1
+- @hyperlane-xyz/provider-sdk@1.3.5
+
+## 25.3.0
+
+### Patch Changes
+
+- @hyperlane-xyz/utils@25.3.0
+- @hyperlane-xyz/provider-sdk@1.3.4
+
 ## 25.2.0
 
 ### Patch Changes